requirements.txt
flask
flask-sqlalchemy
flask-wtf
MySQL-python
flask-paginate
flask-login
flask-admin
flask-babel
flask-script
flask-wtf
flask-mongoengine
Flask-Cache
